size is already size_t.  No need to cast
git-svn-id: https://svn.parrot.org/parrot/trunk@23625 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/gc/memory.c
@@ -46,7 +46,7 @@ PARROT_CANNOT_RETURN_NULL
 void *
 mem_sys_allocate(size_t size)
 {
-    void * const ptr = malloc((size_t)size);
+    void * const ptr = malloc(size);
 #ifdef DETAIL_MEMORY_DEBUG
     fprintf(stderr, "Allocated %i at %p\n", size, ptr);
 #endif
